mcp_meeting_minutes/requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
groq>=0.4.0
pydub>=0.25.0
librosa>=0.10.0